Pull out codegen, allow code-less process definitions
Separated the code generator from the on_build function out into a
modular class. Fix the requires to be a list, rather than a literal
string. Also, allow processes to be generated that don't have any code -
the tlp still contains them, and a requires can pull in the appropriate
module.