Add the default save extension if not present
Under certain conditions the save dialog doesn't append the file
extension to the filename. This causes headaches opening the files
again, as the open dialog is locked to accept only the correct
extension. This is a hacky fix, but the resolution doesn't seem to be in
the current wxPython versions on Mac & Linux.