Propogation and maintainance of generic chanends
Add an attribute to channel ends which stores the generic type, so that
we can have multiple types of generic ends in a process. If one of the
ends in a channel connection being made by the user is generic, that end
takes the type from a specific connection, and specialises all other
generic ends attached to it recursively. Need to implement revertion to
generic type on deletion, and perhaps look at a cleaner way of
implementing this in the YAML.